Re-organize fill-form agent
To ease reading/understanding/debugging of code
