+++ assistants/prospector-assistant/assistant/agents/form_fill_agent/config.py
@@ -2,22 +2,21 @@ from typing import Annotated
 
 from pydantic import BaseModel, Field
 
-from . import gce_config
-from .steps import acquire_form, extract_form_fields, fill_form
+from .steps import acquire_form_step, extract_form_fields_step, fill_form_step, types
 
 
 class FormFillAgentConfig(BaseModel):
     acquire_form_config: Annotated[
-        gce_config.GuidedConversationDefinition,
+        types.GuidedConversationDefinition,
         Field(title="Form Acquisition", description="Guided conversation for acquiring a form from the user."),
-    ] = acquire_form.definition.model_copy()
+    ] = acquire_form_step.definition.model_copy()
 
     extract_form_fields_config: Annotated[
-        extract_form_fields.ExtractFormFieldsConfig,
+        extract_form_fields_step.ExtractFormFieldsConfig,
         Field(title="Extract Form Fields", description="Configuration for extracting form fields from the form."),
-    ] = extract_form_fields.ExtractFormFieldsConfig()
+    ] = extract_form_fields_step.ExtractFormFieldsConfig()
 
     fill_form_config: Annotated[
-        gce_config.GuidedConversationDefinition,
+        types.GuidedConversationDefinition,
         Field(title="Fill Form", description="Guided conversation for filling out the form."),
-    ] = fill_form.definition.model_copy()
+    ] = fill_form_step.definition.model_copy()
